Update index.html
I have no control on how often Github updates all calls to its API for updated stats on download counter. There will inevitably be a delay of 15 mins to several hours. This update to the html markup at least ensures the counter is bypassing whatever agent is being used to access the site, allowing for the correct retrieval of downloads of the counter from the API if any updates exist and are available.